# openapi/main.yaml
openapi: 3.0.3
info:
  title: Pharmafind Authorization Server API
  version: 2.0.0
  description: >
    OpenAPI contract for the Pharmafind custom JWT Authorization Server.
    Provides local registration, login, token refresh, logout and Google OAuth2
    authentication endpoints.
servers:
  - url: http://localhost:9000
tags:
  - name: Authentication
    description: Local account authentication and token management
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequestDTO:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 4
          maxLength: 120
          pattern: '^[a-zA-Z0-9._-]+$'
          example: john.doe
        email:
          type: string
          format: email
          maxLength: 190
          example: john@example.com
        password:
          type: string
          format: password
          example: StrongPassword123!
    RegisterResponseDTO:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
    LoginRequestDTO:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: john.doe
        password:
          type: string
          format: password
          example: StrongPassword123!
    RefreshRequestDTO:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: The refresh token issued during login
    LogoutRequestDTO:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: The refresh token to revoke
    TokenResponseDTO:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          format: int64
    UserInfoResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        provider:
          type: string
        roles:
          type: array
          items:
            type: string
        email_verified:
          type: boolean
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
    ApiErrorResponseDTO:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        violations:
          type: object
          additionalProperties:
            type: string
paths:
  /users:
    get:
      tags:
        - Authentication
      operationId: getAllUsers
      summary: List all users
      description: Returns a non-paginated list of all registered users.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserInfoResponseDTO'
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDTO'
  /user/info:
    get:
      tags:
        - Authentication
      operationId: getCurrentUserInfo
      summary: Get current user information
      description: Returns the authenticated user's profile information based on the JWT token.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponseDTO'
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDTO'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDTO'
  /api/auth/register:
    post:
      tags:
        - Authentication
      operationId: register
      summary: Register a new local user account
      description: Creates a local account stored by the authorization server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDTO'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponseDTO'
        '400':
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDTO'
        '409':
          description: Duplicate username or email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDTO'
  /api/auth/login:
    post:
      tags:
        - Authentication
      operationId: login
      summary: Authenticate with username and password
      description: Validates credentials and returns JWT access and refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDTO'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponseDTO'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDTO'
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      operationId: refreshToken
      summary: Refresh an access token
      description: Exchanges a valid refresh token for a new access token and refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequestDTO'
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponseDTO'
        '400':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDTO'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      operationId: logout
      summary: Revoke a refresh token
      description: Invalidates the given refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequestDTO'
      responses:
        '204':
          description: Token revoked successfully